geolocation/geo-00026.html

<!DOCTYPE html>
<html>
  <head>
    <title>Geolocation Test: window.navigator.geolocation exists and is of type 'object'</title>
    <link rel="author" title="Intel" href="http://www.intel.com/">
    <link rel="help" href="http://www.w3.org/TR/2012/PR-geolocation-API-20120510/#geolocation">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      test(function() {
        assert_true(typeof window.navigator.geolocation == 'object', "window.navigator.geolocation exists and is of type 'object'");
      }, "Test checks that window.navigator.geolocation exists and is of type 'object'");
    </script>
  </body>
</html>
